Made program more easily modifiable
Changed Main method so that it retrieves the Characters to be created from a text file. Changed it so that Character whereabouts are saved within that Character's class instead of in the Setting class. Reduced copy/paste code and made it so that the number of characters can be modified. Moved universal scanner to Main class.